Добавил возможность передавать глюкозу и углеводы в единицах измерения, настроенных больным, не только для записей дневника, но и для профиля пациента (пороговые значения) и запроса на расчет инсулина; + теперь валидация диапазонов значений глюкоз и углеводов зависит от единицы измерения;

+++ src/main/java/com/artlighter/glucosecontrolservice/nightscout/util/mapper/NightscoutTreatmentMapper.java
@@ -53,8 +53,10 @@ public class NightscoutTreatmentMapper implements DTOMapper<List<DiaryEntry>, Ni
     private GlucoseEntry createGlucoseEntry(String commitedAt, Float glucose, String notes, String units) {
         GlucoseEntry glucoseEntry = new GlucoseEntry();
 
-        fill(glucoseEntry, commitedAt,
-                units.equals("mmol") ? glucose : GlucoseUnit.MILLIGRAMS_PER_DECILITER.convertToMmolPerLiter(glucose),
+        fill(glucoseEntry,
+                commitedAt,
+                units.equals("mmol") ? glucose :
+                        (float) GlucoseUnit.MILLIGRAMS_PER_DECILITER.convertToMmolPerLiter(glucose),
                 notes);
 
         return glucoseEntry;
